vps-python-bot/bot/requirements.txt
oandapyV20==0.7.2
transformers==4.33.2
torch==2.0.1
pandas==2.1.0
numpy==1.25.2
python-dotenv==1.0.0
cryptography==41.0.3
passlib==1.7.4
bcrypt==4.0.1
schedule==1.2.0
